ConstraintsQueries move to comon-graphql.
